# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TourVision API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface SuccessResponse {
  success: boolean;
  message?: string;
}

export interface ErrorResponse {
  error: string;
  code?: string;
}

export type UserProfileAccountType =
  (typeof UserProfileAccountType)[keyof typeof UserProfileAccountType];

export const UserProfileAccountType = {
  agent: "agent",
  host: "host",
  developer: "developer",
  manager: "manager",
  other: "other",
} as const;

export type UserProfileSubscriptionTier =
  (typeof UserProfileSubscriptionTier)[keyof typeof UserProfileSubscriptionTier];

export const UserProfileSubscriptionTier = {
  free: "free",
  pro: "pro",
  unlimited: "unlimited",
} as const;

export interface UserProfile {
  id: string;
  fullName: string;
  email: string;
  avatarUrl?: string | null;
  accountType: UserProfileAccountType;
  country?: string | null;
  whatsappNumber?: string | null;
  subscriptionTier: UserProfileSubscriptionTier;
  subscriptionStatus?: string | null;
  toursThisMonth: number;
  totalTours: number;
  onboardingCompleted: boolean;
  createdAt: string;
}

export type UpdateUserProfileBodyAccountType =
  (typeof UpdateUserProfileBodyAccountType)[keyof typeof UpdateUserProfileBodyAccountType];

export const UpdateUserProfileBodyAccountType = {
  agent: "agent",
  host: "host",
  developer: "developer",
  manager: "manager",
  other: "other",
} as const;

export interface UpdateUserProfileBody {
  fullName?: string;
  accountType?: UpdateUserProfileBodyAccountType;
  country?: string;
  whatsappNumber?: string;
}

export type OnboardingBodyAccountType =
  (typeof OnboardingBodyAccountType)[keyof typeof OnboardingBodyAccountType];

export const OnboardingBodyAccountType = {
  agent: "agent",
  host: "host",
  developer: "developer",
  manager: "manager",
  other: "other",
} as const;

export interface OnboardingBody {
  accountType: OnboardingBodyAccountType;
  country?: string;
  whatsappNumber?: string;
}

export interface UserLimits {
  tier: string;
  toursThisMonth: number;
  toursLimit: number;
  toursRemaining: number;
  renewalDate?: string | null;
}

export type TourStatusProperty =
  (typeof TourStatusProperty)[keyof typeof TourStatusProperty];

export const TourStatusProperty = {
  pending: "pending",
  extracting: "extracting",
  generating: "generating",
  stitching: "stitching",
  building_3d: "building_3d",
  ready: "ready",
  failed: "failed",
} as const;

export type TourMarbleWorldIds = { [key: string]: unknown } | null;

export interface Tour {
  id: string;
  userId: string;
  listingUrl: string;
  listingTitle?: string | null;
  listingAddress?: string | null;
  listingPlatform?: string | null;
  listingPrice?: string | null;
  listingBedrooms?: number | null;
  listingBathrooms?: number | null;
  listingSqft?: string | null;
  status: TourStatusProperty;
  currentStage?: string | null;
  totalPhotosExtracted: number;
  photosUsed: number;
  roomsDetected: number;
  floorCount?: number | null;
  confidenceScore: number;
  realAngles: number;
  aiHighAngles: number;
  aiLowAngles: number;
  marbleWorldIds?: TourMarbleWorldIds;
  tourEmbedUrl?: string | null;
  shareToken?: string | null;
  isWatermarked: boolean;
  thumbnailUrl?: string | null;
  viewCount: number;
  errorMessage?: string | null;
  processingStartedAt?: string | null;
  processingCompletedAt?: string | null;
  createdAt: string;
}

export interface TourRoom {
  id: string;
  tourId: string;
  roomLabel: string;
  floorNumber?: number | null;
  qualityScore: number;
  isSelected: boolean;
  isBestForRoom: boolean;
  confidenceScore?: number | null;
  marbleWorldId?: string | null;
  marbleEmbedUrl?: string | null;
  thumbnailUrl?: string | null;
  isAiGenerated: boolean;
  createdAt: string;
}

export interface TourRoomsResponse {
  rooms: TourRoom[];
  floorCount?: number | null;
}

export type UpdateRoomFloorsBodyAssignmentsItem = {
  roomId: string;
  floorNumber: number;
};

export interface UpdateRoomFloorsBody {
  assignments: UpdateRoomFloorsBodyAssignmentsItem[];
}

export interface SetFloorCountBody {
  floorCount: number;
}

export interface ToursListResponse {
  tours: Tour[];
  total: number;
  page?: number;
  totalPages?: number;
}

export interface CreateTourBody {
  listingUrl: string;
  floorCount?: number;
}

export interface CreateTourResponse {
  tourId: string;
  shareToken: string;
}

export interface TourStatus {
  status: string;
  currentStage?: string | null;
  roomsCompleted: number;
  roomsTotal: number;
  estimatedMinutes: number;
  confidenceScore: number;
  errorMessage?: string | null;
}

export interface TourStats {
  toursThisMonth: number;
  toursLimit: number;
  avgProcessingMinutes: number;
  totalViewsThisMonth: number;
  totalToursAllTime: number;
}

export type PublicTourMarbleWorldIds = { [key: string]: unknown } | null;

export interface PublicTour {
  id: string;
  shareToken: string;
  listingAddress?: string | null;
  listingPlatform?: string | null;
  listingPrice?: string | null;
  listingBedrooms?: number | null;
  listingBathrooms?: number | null;
  listingSqft?: string | null;
  marbleWorldIds?: PublicTourMarbleWorldIds;
  rooms: TourRoom[];
  confidenceScore: number;
  realAngles: number;
  aiHighAngles: number;
  aiLowAngles: number;
  isWatermarked: boolean;
  agentName?: string | null;
  agentLogo?: string | null;
  thumbnailUrl?: string | null;
}

export interface BuyerLeadBody {
  name: string;
  email: string;
  phone?: string;
  message?: string;
}

export type BuyerLeadStatus =
  (typeof BuyerLeadStatus)[keyof typeof BuyerLeadStatus];

export const BuyerLeadStatus = {
  new: "new",
  contacted: "contacted",
  converted: "converted",
} as const;

export interface BuyerLead {
  id: string;
  tourId: string;
  buyerName: string;
  buyerEmail: string;
  buyerPhone?: string | null;
  message?: string | null;
  status: BuyerLeadStatus;
  tourAddress?: string | null;
  createdAt: string;
}

export interface LeadsListResponse {
  leads: BuyerLead[];
  total: number;
}

export type AnalyticsOverviewToursOverTimeItem = {
  date: string;
  count: number;
};

export type AnalyticsOverviewViewsByTourItem = {
  tourId: string;
  address: string;
  views: number;
};

export type AnalyticsOverviewViewsByCountryItem = {
  country: string;
  views: number;
  percentage: number;
};

export type AnalyticsOverviewProcessingTrendItem = {
  date: string;
  avgMinutes: number;
};

export type AnalyticsOverviewTourPerformanceItem = {
  tourId: string;
  address: string;
  views: number;
  countries: number;
  avgTimeInTour: number;
  leads: number;
};

export interface AnalyticsOverview {
  toursOverTime: AnalyticsOverviewToursOverTimeItem[];
  viewsByTour: AnalyticsOverviewViewsByTourItem[];
  viewsByCountry: AnalyticsOverviewViewsByCountryItem[];
  processingTrend: AnalyticsOverviewProcessingTrendItem[];
  tourPerformance: AnalyticsOverviewTourPerformanceItem[];
}

export interface FlagAngleBody {
  tourId: string;
  angleId: string;
  reason?: string;
}

export interface NewsletterSubscribeBody {
  email: string;
}

export type ListToursParams = {
  status?: ListToursStatus;
  search?: string;
  sort?: ListToursSort;
  page?: number;
  limit?: number;
};

export type ListToursStatus =
  (typeof ListToursStatus)[keyof typeof ListToursStatus];

export const ListToursStatus = {
  all: "all",
  ready: "ready",
  processing: "processing",
  failed: "failed",
} as const;

export type ListToursSort = (typeof ListToursSort)[keyof typeof ListToursSort];

export const ListToursSort = {
  newest: "newest",
  oldest: "oldest",
  confidence: "confidence",
  views: "views",
} as const;