# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: TourVision API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /tours/{tourId}/lead:
    post:
      operationId: submitBuyerLead
      tags: [leads]
      summary: Submit buyer lead from tour viewer (public)
      parameters:
        - name: tourId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BuyerLeadBody"
      responses:
        "200":
          description: Lead submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
  /angles/flag:
    post:
      operationId: flagAngle
      tags: [flags]
      summary: Flag an angle as incorrect (public)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FlagAngleBody"
      responses:
        "200":
          description: Flag submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
...
    BuyerLeadBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        message:
          type: string
      required:
        - name
        - email
...
    FlagAngleBody:
      type: object
      properties:
        tourId:
          type: string
        angleId:
          type: string
        reason:
          type: string
      required:
        - tourId
        - angleId
...